requirements.txt
asgiref==3.8.1
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.15.2
gunicorn==23.0.0
uvicorn==0.15.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
